backend/requirements/base.txt
django==4.2.*
djangorestframework==3.14.*
psycopg2-binary==2.9.*
redis==4.5.*
django-redis==5.2.*
celery==5.3.*
djangorestframework-simplejwt==5.2.*
pillow==10.0.*
django-cors-headers==4.2.*
channels==4.0.*
gunicorn==21.2.*
python-decouple==3.8.*
django-extensions==3.2.*
whitenoise==6.5.*
django-filter==23.2.*
dj-database-url==2.1.0
python-magic==0.4.27
pyclamd==0.4.0
django-imagekit==5.0.0
sentry-sdk[django]==1.40.6
django-encrypted-model-fields==0.6.5
django-ninja==1.0.*